updates to tractor beam, prospector, fleetID code, bugfix #56 bugfix #71

+++ src/eve-server/system/SystemManager.h
@@ -112,7 +112,7 @@ public:
     // range is 0.1 for 1.0 system to 2.0 for -0.9 system
     float GetSecValue()                                 { return m_secValue; }
 
-    bool BuildDynamicEntity(const DBSystemDynamicEntity& entity, int64 launcherID=0);
+    bool BuildDynamicEntity(const DBSystemDynamicEntity& entity, uint32 launcherID=0);
 
     void AddNPC(NPC* who);
     void RemoveNPC(NPC* who);
